% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parl.R
\name{parl_member}
\alias{parl_member}
\title{Get a specific UK Parliament member by ID}
\usage{
parl_member(id)
}
\arguments{
\item{id}{Member ID (integer)}
}
\value{
tibble: id, name, full_title, party, constituency, house,
gender, start_date, status
}
\description{
Get a specific UK Parliament member by ID
}
