% Generated by build script
\name{parl_member}
\alias{parl_member}
\title{Get a specific UK Parliament member by ID}
\usage{
parl_member()
}
\arguments{
\item{id}{Member ID (integer)}
}
\value{tibble: id, name, full_title, party, constituency, house,}
\description{Get a specific UK Parliament member by ID}
